+++ src/reset-scoped/react/index.ts
@@ -2,11 +2,10 @@ import * as React from "react";
 
 import { transformProps } from "./transformProps";
 
-/**
- * @param {React.ElementType} type
- * @param {Object.<string, *>} props
- * @param {...*} restArgs
- */
-export function createElement(type, props, ...restArgs) {
+export function createElement(
+  type: React.ElementType,
+  props: { [key: string]: any },
+  ...restArgs: any[]
+) {
   return React.createElement(type, transformProps(type, props), ...restArgs);
 }