+++ src/reset-scoped/react/jsx-dev-runtime.ts
@@ -2,14 +2,15 @@ import * as ReactJSXDEVRuntime from "react/jsx-dev-runtime";
 
 import { transformProps } from "./transformProps";
 
+// @ts-expect-error
 export const { Fragment } = ReactJSXDEVRuntime;
 
-/**
- * @param {React.ElementType} type
- * @param {Object.<string, *>} props
- * @param {...*} restArgs
- */
-export function jsxDEV(type, props, ...restArgs) {
+export function jsxDEV(
+  type: React.ElementType,
+  props: { [key: string]: any },
+  ...restArgs: any[]
+) {
+  // @ts-expect-error
   return ReactJSXDEVRuntime.jsxDEV(
     type,
     transformProps(type, props),