+++ src/reset-scoped/react/jsx-runtime.ts
@@ -2,22 +2,23 @@ import * as ReactJSXRuntime from "react/jsx-runtime";
 
 import { transformProps } from "./transformProps";
 
+// @ts-expect-error
 export const { Fragment } = ReactJSXRuntime;
 
-/**
- * @param {React.ElementType} type
- * @param {Object.<string, *>} props
- * @param {...*} restArgs
- */
-export function jsx(type, props, ...restArgs) {
+export function jsx(
+  type: React.ElementType,
+  props: { [key: string]: any },
+  ...restArgs: any[]
+) {
+  // @ts-expect-error
   return ReactJSXRuntime.jsx(type, transformProps(type, props), ...restArgs);
 }
 
-/**
- * @param {React.ElementType} type
- * @param {Object.<string, *>} props
- * @param {...*} restArgs
- */
-export function jsxs(type, props, ...restArgs) {
+export function jsxs(
+  type: React.ElementType,
+  props: { [key: string]: any },
+  ...restArgs: any[]
+) {
+  // @ts-expect-error
   return ReactJSXRuntime.jsxs(type, transformProps(type, props), ...restArgs);
 }