+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- **reset-scoped:** ship CommonJS modules alongside ESM
+
 ## [3.1.0] - 2023-06-04
 
 ### Added